better textures
dumb java integer overriding thingy mabober....

+++ source/naftoreiclag/village/MapData.java
@@ -40,7 +40,7 @@ public class MapData
 		{
 			for(int z = 0; z < size; ++ z)
 			{
-				geo.put(x - 16).put(value[x][z] - 2.0f).put(z - 16).put(x / size).put(z / size);
+				geo.put(x - 16).put(value[x][z] - 2.0f).put(z - 16).put(x / ((float) size)).put(z / ((float) size));
 			}
 		}
